version 1.0.0

+++ client/src/components/Equipments/CreateEquipment.vue
@@ -1,68 +1,66 @@
 <template>
   <div class="container book-wrapper">
     <main-header navsel="back"></main-header>
-    <h1>Edit Book</h1>
-    <form v-on:submit.prevent = "editBook" >
-      <!-- <p>title: <input type="text" v-model="book.title"></p> -->
+    <h1>เพิ่มอุปกรณ์</h1>
+    <form v-on:submit.prevent = "createEquipment">
       <p>
-        <label for="" class="control-label">Title: </label>
-        <input type="text" v-model="book.title" class="form-control">        
+        <label for="" class="control-label">ชื่อออุปกรณ์: </label>
+        <input type="text" v-model="equipment.title" class="form-control">        
       </p>
       <transition name="fade">
-        <div class="thumbnail-pic" v-if="book.thumbnail != 'null'">
-          <img class="img-thumbnail" :src="BASE_URL+book.thumbnail" alt="thumbnail">
+        <div class="thumbnail-pic" v-if="equipment.thumbnail != 'null'">
+          <img class="img-thumbnail" :src="BASE_URL+equipment.thumbnail" alt="thumbnail">
         </div>
-      </transition>  
-      <form id="upload-form" enctype="multipart/form-data" novalidate>
+      </transition>
+      <form enctype="multipart/form-data" novalidate>
         <div class="dropbox">
           <input type="file" multiple :name="uploadFieldName" :disabled="isSaving" @change="filesChange($event.target.name, $event.target.files); fileCount = $event.target.files.length"
             accept="image/*" class="input-file">
             <!-- <p v-if="isInitial || isSuccess"> -->
             <p v-if="isInitial">
-              Drag your file(s) here to begin<br> or click to browse
+              ลากไฟล์ของคุณมาที่นี่เพื่อเริ่มต้น<br> หรือคลิกเพื่อเรียกดู
             </p>
             <p v-if="isSaving">
-              Uploading {{ fileCount }} files...
+              กำลังอัปโหลด {{ fileCount }} ไฟล์...
             </p>   
             <p v-if="isSuccess">
-              Upload Successful.
-            </p>        
+              อัปโหลดสำเร็จ
+            </p>           
         </div>
       </form>
       <div>
-      <transition-group tag="ul" name="fade" class="pictures">        
-        <li v-for="picture in pictures" v-bind:key="picture.id">              
-          <img class="img-thumbnail" style="margin-bottom:5px;" :src="BASE_URL+picture.name" alt="picture image">              
-          <br />  
-          <button class="btn btn-xs btn-info" v-on:click.prevent="useThumbnail(picture.name)">Thumbnail</button>
-          <button class="btn btn-xs btn-danger" v-on:click.prevent="delFile(picture)">Delete</button>
-        </li>        
-      </transition-group>
-      <div class="clearfix"></div>
+        <transition-group tag="ul" name="fade" class="pictures">        
+          <li v-for="picture in pictures" v-bind:key="picture.id">              
+            <img class="img-thumbnail" style="margin-bottom:5px;" :src="BASE_URL+picture.name" alt="picture image">              
+            <br />  
+            <button class="btn btn-xs btn-info" v-on:click.prevent="useThumbnail(picture.name)">ภาพขนาดย่อ</button>
+            <button class="btn btn-xs btn-danger" v-on:click.prevent="delFile(picture)">ลบ</button>
+          </li>        
+        </transition-group>
+        <div class="clearfix"></div>
       </div>  
-      <p><strong>content: </strong></p>
-      <p><vue-ckeditor v-model.lazy="book.content" :config="config" @blur="onBlur($event)" @focus="onFocus($event)" /></p>            
+      <p><strong>เนื้อหา: </strong></p>
+      <p><vue-ckeditor v-model.lazy="equipment.content" :config="config" @blur="onBlur($event)" @focus="onFocus($event)" /></p>
       <p>
-        <label class="control-label">Category :</label>
-        <input type="text" v-model="book.category" class="form-control">
-      </p>     
+        <label class="control-label">ประเภท :</label>
+        <input type="text" v-model="equipment.category" class="form-control">
+      </p> 
       <p>
-        <label class="control-label">Prices :</label>
-        <input type="text" v-model="book.prices" class="form-control">
-      </p>       
+        <label class="control-label">ราคาเมื่ออุปกร์เสียหาย :</label>
+        <input type="text" v-model="equipment.prices" class="form-control">
+      </p> 
       <p>
-        <button class="btn btn-warning" type="submit">Update Book</button>
-        <button class="btn btn-default" type="button" v-on:click="navigateTo('/books')">Back</button>
-      </p>      
-      
+        <button class="btn btn-success" type="submit">เพิ่มอุปกรณ์</button>
+        <button class="btn btn-default" type="button" v-on:click="navigateTo('/equipments')">กลับ</button>
+      </p> 
     </form>   
     <br>    
   </div>
 </template>
 <script>
-import BooksService from '@/services/BooksService'
-import UploadService from '@/services/UploadService'
+import EquipmentsService from '@/services/EquipmentsService'
 import VueCkeditor from "vue-ckeditor2"
+import UploadService from '@/services/UploadService'
 import {mapState} from 'vuex'
 
 const STATUS_INITIAL = 0,
@@ -71,27 +69,30 @@ const STATUS_INITIAL = 0,
   STATUS_FAILED = 3
 
 export default {
+  mounted () {
+    if (!this.isUserLoggedIn) {
+      this.$router.push({
+        name: 'login'        
+      })
+    }
+  },
   data () {
     return {
-      // book data
-      book: {
+      equipment: {
         title: '',
         thumbnail: 'null',
-        pictures: 'null',
+        pictures: [],
         content: '',
         category: '',
-        status: '',
+        status: 'Suspend',
         prices: ''
       },
-
-      // ckeditor
       config: {
         toolbar: [
           ["Bold", "Italic", "Underline", "Strike", "Subscript", "Superscript"]
         ],
         height: 300
       },
-
       // upload data
       BASE_URL: "http://localhost:8081/assets/uploads/",
       error: null,      
@@ -99,59 +100,36 @@ export default {
       currentStatus: null,
       uploadFieldName: "userPhoto",      
       uploadedFileNames:[],
-      pictureIndex: 0,
-
-      // show result
       pictures: [],
+      pictureIndex: 0,
     }
   },
-  methods: {
+  methods: {// upload
     navigateTo (route) {
       this.$router.push(route)
     },
-    // ckedior ()
-    onBlur(editor) {
-      console.log(editor);
-    },
-    onFocus(editor) {
-      console.log(editor);
-    },
-    
-    // book
-    async editBook () {
-      this.book.pictures = JSON.stringify(this.pictures)
-      try {
-        await BooksService.put(this.book)
-        this.$router.push({
-          name: 'books'
-        })
-      } catch (err) {
-        console.log(err)
-      }
-    },
-
-    // manage pic field
     useThumbnail (filename) {     
       console.log(filename) 
-      this.book.thumbnail = filename
+      this.equipment.thumbnail = filename
     },
     async delFile (material){
-      let dataJSON = {
-        "filename":material.name
-      }
-      
-      await UploadService.delete(dataJSON)      
-
-      for (var i=0;i<this.pictures.length;i++){
-        if(this.pictures[i].id === material.id) {
-          this.pictures.splice(i, 1)
-          this.materialIndex--
-          break
+      let result = confirm("Want to delete?")
+      if (result) {
+        let dataJSON = {
+          "filename":material.name
         }
-      }     
-    },
+        
+        await UploadService.delete(dataJSON)      
 
-    // upload
+        for (var i=0;i<this.pictures.length;i++){
+          if(this.pictures[i].id === material.id) {
+            this.pictures.splice(i, 1)
+            this.materialIndex--
+            break
+          }
+        }    
+      } 
+    },
     wait(ms) {
       return x => {
         return new Promise(resolve => setTimeout(() => resolve(x), ms));
@@ -192,8 +170,6 @@ export default {
           }  
         })
 
-
-        this.clearUploadResult()
       } catch (error) {
         console.log(error)
         this.currentStatus = STATUS_FAILED
@@ -212,16 +188,32 @@ export default {
       });
 
       // save it
-      this.save(formData)
-
-      // clear form
-      document.getElementById("upload-form").reset()
+      this.save(formData);
     },
     clearUploadResult: function(){            
       setTimeout(() => this.reset(), 5000);
-    }
+    },
+    async createEquipment () {
+      this.equipment.pictures = JSON.stringify(this.pictures)
+      try {
+        await EquipmentsService.post(this.equipment)
+        this.$router.push({
+          name: 'equipments'
+        })
+      } catch (err) {
+        console.log(err)
+      }
+    },
+    onBlur (editor) {
+      console.log(editor);
+    },
+    onFocus (editor) {
+      console.log(editor);
+    },
   },
-
+  components: { 
+    VueCkeditor 
+  },
   computed: {
     ...mapState([
       'isUserLoggedIn',
@@ -240,27 +232,8 @@ export default {
       return this.currentStatus === STATUS_FAILED;
     }
   },
-
-  async mounted () {
-    if (!this.isUserLoggedIn) {
-      this.$router.push({
-        name: 'login'        
-      })
-    }
-
-    try {      
-      let bookId = this.$route.params.bookId
-      this.book = (await BooksService.show(bookId)).data 
-      this.pictures = JSON.parse(this.book.pictures)
-    } catch (error) {
-      console.log (error)
-    }
-  },
-  
-  created () {    
-
-    this.reset()
-
+  created () {
+    this.reset(),
     this.config.toolbar = [
       {
         name: "document",
@@ -363,11 +336,8 @@ export default {
       { name: "colors", items: ["TextColor", "BGColor"] },
       { name: "tools", items: ["Maximize", "ShowBlocks"] },
       { name: "about", items: ["About"] }
-    ] 
-  },
-  components: { 
-    VueCkeditor 
-  },
+    ]
+  }
 }
 </script>
 <style scoped>
@@ -379,24 +349,6 @@ export default {
 .thumbnail-pic img{
   width:200px;
   margin-bottom: 10px;
-}
-/* display uploaded pic */
-ul.pictures {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  float: left;
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
-ul.pictures li {
-  float: left;
-}
-
-ul.pictures li img {
-  max-width: 180px;
-  margin-right: 20px;
 }
 
 /* uplaod */
@@ -427,5 +379,27 @@ ul.pictures li img {
   font-size: 1.2em;
   text-align: center;
   padding: 50px 0;
+}
+
+/* display uplaod */
+ul.pictures {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  float: left;
+  padding-top: 10px;
+  padding-bottom: 10px;
+}
+
+ul.pictures li {
+  float: left;
+}
+
+ul.pictures li img {
+  max-width: 180px;
+  margin-right: 20px;
+}
+div {
+    font-family: 'Kanit', sans-serif;
 }
 </style>
